Added CrossSplitter, gloabl replace double->float, reverted unnecessary LINQ expressions that created temporary collections, minor code cleanup.

+++ GwenCS/GwenCS/Anim/Size/Height.cs
@@ -11,7 +11,7 @@ namespace Gwen.Anim.Size
         protected int m_iDelta;
         protected bool m_bHide;
 
-        public Height(int iStartSize, int iEndSize, double fLength, bool bHide = false, double fDelay = 0.0, double fEase = 1.0) 
+        public Height(int iStartSize, int iEndSize, float fLength, bool bHide = false, float fDelay = 0.0f, float fEase = 1.0f) 
             : base( fLength, fDelay, fEase )
         {
             m_iStartSize = iStartSize;
@@ -20,7 +20,7 @@ namespace Gwen.Anim.Size
         }
 
         protected override void onStart() { base.onStart(); m_Control.Height = m_iStartSize; }
-        protected override void Run(double delta) { base.Run(delta); m_Control.Height = Global.Trunc( m_iStartSize + (m_iDelta * delta)); }
+        protected override void Run(float delta) { base.Run(delta); m_Control.Height = Global.Trunc( m_iStartSize + (m_iDelta * delta)); }
         protected override void onFinish() { base.onFinish(); m_Control.Height = m_iStartSize + m_iDelta; m_Control.IsHidden = m_bHide; }
     }
 }
