+++ GwenCS/GwenCS/Anim/TimedAnimation.cs
@@ -7,13 +7,13 @@ namespace Gwen.Anim
     {
         protected bool m_bStarted;
         protected bool m_bFinished;
-        protected double m_fStart;
-        protected double m_fEnd;
-        protected double m_fEase;
+        protected float m_fStart;
+        protected float m_fEnd;
+        protected float m_fEase;
 
         public override bool Finished { get { return m_bFinished; } }
 
-        public TimedAnimation( double fLength, double fDelay = 0.0, double fEase = 1.0 )
+        public TimedAnimation( float fLength, float fDelay = 0.0f, float fEase = 1.0f )
         {
             m_fStart = Platform.Windows.GetTimeInSeconds() + fDelay;
             m_fEnd = m_fStart + fLength;
@@ -29,8 +29,8 @@ namespace Gwen.Anim
             if (m_bFinished) 
                 return;
 
-            double fCurrent = Platform.Windows.GetTimeInSeconds();
-            double fSecondsIn = fCurrent - m_fStart;
+            float fCurrent = Platform.Windows.GetTimeInSeconds();
+            float fSecondsIn = fCurrent - m_fStart;
             if (fSecondsIn < 0.0) 
                 return;
 
@@ -40,15 +40,15 @@ namespace Gwen.Anim
                 onStart();
             }
 
-            double fDelta = fSecondsIn/(m_fEnd - m_fStart);
-            if (fDelta < 0.0) 
-                fDelta = 0.0;
-            if (fDelta > 1.0) 
-                fDelta = 1.0;
+            float fDelta = fSecondsIn/(m_fEnd - m_fStart);
+            if (fDelta < 0.0f) 
+                fDelta = 0.0f;
+            if (fDelta > 1.0f) 
+                fDelta = 1.0f;
 
-            Run(Math.Pow(fDelta, m_fEase));
+            Run((float)Math.Pow(fDelta, m_fEase));
 
-            if (fDelta == 1.0)
+            if (fDelta == 1.0f)
             {
                 m_bFinished = true;
                 onFinish();
@@ -60,7 +60,7 @@ namespace Gwen.Anim
         protected virtual void onStart()
         {}
 
-        protected virtual void Run(double delta)
+        protected virtual void Run(float delta)
         {}
 
         protected virtual void onFinish()
