+++ GwenCS/GwenCS/Input/KeyData.cs
@@ -6,7 +6,7 @@ namespace Gwen.Input
     public class KeyData
     {
         public bool[] KeyState;
-        public double[] NextRepeat;
+        public float [] NextRepeat;
         public Base Target;
         public bool LeftMouseDown;
         public bool RightMouseDown;
@@ -14,7 +14,7 @@ namespace Gwen.Input
         public KeyData()
         {
             KeyState = new bool[(int)Key.Count];
-            NextRepeat = new double[(int)Key.Count];
+            NextRepeat = new float[(int)Key.Count];
             // everything is initialized to 0 by default
         }
     }
